refactor: extend from base/update migrations
